update from new evcc-io

+++ core/loadpoint_adapter.go
@@ -1,6 +1,6 @@
 package core
 
-import "github.com/andig/evcc/core/soc"
+import "github.com/evcc-io/evcc/core/soc"
 
 type adapter struct {
 	lp *LoadPoint
@@ -18,7 +18,7 @@ func (a *adapter) SocEstimator() *soc.Estimator {
 	return a.lp.socEstimator
 }
 
-func (a *adapter) ActivePhases() int64 {
+func (a *adapter) ActivePhases() int {
 	return a.lp.activePhases
 }
